// src/generated/client/src/models/Post.ts
export interface Post {
    /**
     * The date the object was published, in the site\'s timezone.
     * @type {Date}
     * @memberof Post
     */
    date?: Date;
    /**
     * The date the object was published, as GMT.
     * @type {Date}
     * @memberof Post
     */
    date_gmt?: Date;
    /**
     * 
     * @type {PostGuid}
     * @memberof Post
     */
    guid?: PostGuid;
    /**
     * Unique identifier for the object.
     * @type {number}
     * @memberof Post
     */
    id?: number;
    /**
     * URL to the object.
     * @type {string}
     * @memberof Post
     */
    link?: string;
    /**
     * The date the object was last modified, in the site\'s timezone.
     * @type {Date}
     * @memberof Post
     */
    modified?: Date;
    /**
     * The date the object was last modified, as GMT.
     * @type {Date}
     * @memberof Post
     */
    modified_gmt?: Date;
    /**
     * An alphanumeric identifier for the object unique to its type.
     * @type {string}
     * @memberof Post
     */
    slug?: string;
    /**
     * A named status for the object.
     * @type {string}
     * @memberof Post
     */
    status?: PostStatusEnum;
    /**
     * Type of Post for the object.
     * @type {string}
     * @memberof Post
     */
    type?: string;
    /**
     * A password to protect access to the content and excerpt.
     * @type {string}
     * @memberof Post
     */
    password?: string;
    /**
     * Permalink template for the object.
     * @type {string}
     * @memberof Post
     */
    permalink_template?: string;
    /**
     * Slug automatically generated from the object title.
     * @type {string}
     * @memberof Post
     */
    generated_slug?: string;
    /**
     * 
     * @type {PostTitle}
     * @memberof Post
     */
    title?: PostTitle;
    /**
     * 
     * @type {PostContent}
     * @memberof Post
     */
    content?: PostContent;
    /**
     * The ID for the author of the object.
     * @type {number}
     * @memberof Post
     */
    author?: number;
    /**
     * 
     * @type {PostExcerpt}
     * @memberof Post
     */
    excerpt?: PostExcerpt;
    /**
     * The ID of the featured media for the object.
     * @type {number}
     * @memberof Post
     */
    featured_media?: number;
    /**
     * Whether or not comments are open on the object.
     * @type {string}
     * @memberof Post
     */
    comment_status?: PostCommentStatusEnum;
    /**
     * Whether or not the object can be pinged.
     * @type {string}
     * @memberof Post
     */
    ping_status?: PostPingStatusEnum;
    /**
     * The format for the object.
     * @type {string}
     * @memberof Post
     */
    format?: PostFormatEnum;
    /**
     * Meta fields.
     * @type {object}
     * @memberof Post
     */
    meta?: object;
    /**
     * Whether or not the object should be treated as sticky.
     * @type {boolean}
     * @memberof Post
     */
    sticky?: boolean;
    /**
     * The theme file to use to display the object.
     * @type {string}
     * @memberof Post
     */
    template?: string;
    /**
     * The terms assigned to the object in the category taxonomy.
     * @type {Array<number>}
     * @memberof Post
     */
    categories?: Array<number>;
    /**
     * The terms assigned to the object in the post_tag taxonomy.
     * @type {Array<number>}
     * @memberof Post
     */
    tags?: Array<number>;
}
// ...
export enum PostStatusEnum {
    Publish = 'publish',
    Future = 'future',
    Draft = 'draft',
    Pending = 'pending',
    Private = 'private'
}
/**
* @export
* @enum {string}
*/
export enum PostCommentStatusEnum {
    Open = 'open',
    Closed = 'closed'
}
/**
* @export
* @enum {string}
*/
export enum PostPingStatusEnum {
    Open = 'open',
    Closed = 'closed'
}
/**
* @export
* @enum {string}
*/
export enum PostFormatEnum {
    Standard = 'standard',
    Aside = 'aside',
    Chat = 'chat',
    Gallery = 'gallery',
    Link = 'link',
    Image = 'image',
    Quote = 'quote',
    Status = 'status',
    Video = 'video',
    Audio = 'audio'
}